AAO/PINN/app/requirements.txt
numpy==1.24.3
pandas==2.0.1
scipy==1.10.1
matplotlib==3.7.1
optax==0.1.5
dm-haiku==0.0.9
typing-extensions==4.5.0